myapp/requirements.txt
asgiref==3.8.1
babel==2.16.0
casbin==1.36.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==0.61.1
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-money==3.0.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-aggregation==1.1.2
drf-complex-filter==1.3.0
drf-flex-fields==1.0.2
drf-yasg==1.21.7
filelock==3.15.4
idna==3.8
inflection==0.5.1
Markdown==3.7
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
py-moneyed==2.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
simpleeval==0.9.13
sorl-thumbnail==12.10.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.26.3


django
djangorestframework
environs
mysqlclient
site-packages (2024.1)
site-packages (0.5.1)